The emulated Undo_stack becomes more stable
Catches some corner cases of undo_stack: whitespaces and newlines in the lyrics broke undo_stack before. Those have been fixed now.
Also fix the bugs on behaviors of prevChar() and nextChar(). when they reaches the filehead or tail and loop the lrcCursor, if the filehead/tail is a whitespace or \n, it selects the whitespace.
Also add the UI for some unfinished features.